adding IndependentTrustee types: context | provider | i18n
