+++ packages/core/src/types/common/providers/index.ts
@@ -2,3 +2,4 @@ export * from './employer';
 export * from './trustee';
 export * from './actuary';
 export * from './corporateGroup';
+export * from './independentTrustee';
