+++ packages/layout/src/components/__tests__/corporateGroup.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { CorporateGroupCard } from '../cards/corporateGroup/corporateGroup';
-import { CorporateGroup } from '../cards/corporateGroup/context';
+import { CorporateGroup } from '@tpr/core';
 import { axe } from 'jest-axe';
 import { cleanup } from '@testing-library/react-hooks';
 import { act } from 'react-dom/test-utils';
