+++ packages/layout/src/components/cards/independentTrustee/i18n.tsx
@@ -1,47 +1,4 @@
-import { I18nRemoveReason } from '../common/interfaces';
-
-export type IndependentTrusteeI18nProps = {
-	preview: {
-		buttons: {
-			one: string;
-			two: string;
-			three: string;
-			four: string;
-		};
-		checkboxLabel: string;
-		trusteeType: string;
-	};
-	regulator: {
-		title: string;
-		subtitle: string;
-		fields: {
-			appointedByRegulator: {
-				labels: {
-					isAppointedByRegulatorYes: string;
-					isAppointedByRegulatorNo: string;
-				};
-			};
-		};
-	};
-	remove: {
-		confirm: {
-			title: string;
-			subtitle: string;
-			breadcrumbs: {
-				link1: string;
-				link2: string;
-			};
-			dialog: {
-				message1: string;
-			};
-			buttons: {
-				remove: string;
-				cancel: string;
-			};
-		};
-		reason: I18nRemoveReason;
-	};
-};
+import { IndependentTrusteeI18nProps } from '@tpr/core';
 
 export const i18n: IndependentTrusteeI18nProps = {
 	preview: {
